+++ example/lib/modals/floating_modal.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+
+class FloatingModal extends StatelessWidget {
+  final Widget child;
+  final Color backgroundColor;
+
+  const FloatingModal({Key key, this.child, this.backgroundColor})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 20),
+        child: Material(
+          color: backgroundColor,
+          clipBehavior: Clip.antiAlias,
+          borderRadius: BorderRadius.circular(12),
+          child: child,
+        ),
+      ),
+    );
+  }
+}
+
+Future<T> showFloatingModalBottomSheet<T>({
+  @required BuildContext context,
+  @required ScrollWidgetBuilder builder,
+  Color backgroundColor,
+}) async {
+  final result = await showCustomModalBottomSheet(
+      context: context,
+      builder: builder,
+      containerWidget: (_, animation, child) => FloatingModal(
+            child: child,
+          ),
+      expand: false);
+
+  return result;
+}